# src/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WP07 Local Module - OpenAPI 3.0",
    "description": "",
    "version": "0.2.0"
  },
  "paths": {
    "/digest/pdf": {
      "post": {
        "summary": "Returns digest to be signed of an unsigned PDF",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "bytes",
                  "signingTimestamp"
                ],
                "type": "object",
                "properties": {
                  "bytes": {
                    "type": "string",
                    "description": "Base64 encoded PDF",
                    "example": "JVBERi0xLjEKJcKlwrHDqwoKMSAwIG9iagogIDw8IC9UeXBlIC9DYXRhbG9nCiAgICAgL1BhZ2VzIDIgMCBSCiAgPj4KZW5kb2JqCgoyIDAgb2JqCiAgPDwgL1R5cGUgL1BhZ2VzCiAgICAgL0tpZHMgWzMgMCBSXQogICAgIC9Db3VudCAxCiAgICAgL01lZGlhQm94IFswIDAgMzAwIDE0NF0KICA+PgplbmRvYmoKCjMgMCBvYmoKICA8PCAgL1R5cGUgL1BhZ2UKICAgICAgL1BhcmVudCAyIDAgUgogICAgICAvUmVzb3VyY2VzCiAgICAgICA8PCAvRm9udAogICAgICAgICAgIDw8IC9GMQogICAgICAgICAgICAgICA8PCAvVHlwZSAvRm9udAogICAgICAgICAgICAgICAgICAvU3VidHlwZSAvVHlwZTEKICAgICAgICAgICAgICAgICAgL0Jhc2VGb250IC9UaW1lcy1Sb21hbgogICAgICAgICAgICAgICA+PgogICAgICAgICAgID4+CiAgICAgICA+PgogICAgICAvQ29udGVudHMgNCAwIFIKICA+PgplbmRvYmoKCjQgMCBvYmoKICA8PCAvTGVuZ3RoIDU1ID4+CnN0cmVhbQogIEJUCiAgICAvRjEgMTggVGYKICAgIDAgMCBUZAogICAgKEhlbGxvIFdvcmxkKSBUagogIEVUCmVuZHN0cmVhbQplbmRvYmoKCnhyZWYKMCA1CjAwMDAwMDAwMDAgNjU1MzUgZiAKMDAwMDAwMDAxOCAwMDAwMCBuIAowMDAwMDAwMDc3IDAwMDAwIG4gCjAwMDAwMDAxNzggMDAwMDAgbiAKMDAwMDAwMDQ1NyAwMDAwMCBuIAp0cmFpbGVyCiAgPDwgIC9Sb290IDEgMCBSCiAgICAgIC9TaXplIDUKICA+PgpzdGFydHhyZWYKNTY1CiUlRU9GCg=="
                  },
                  "signingTimestamp": {
                    "type": "number",
                    "description": "Timestamp in milliseconds since epoch",
                    "example": 1676906127623
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bytes": {
                      "type": "string",
                      "description": "Base64 encoded SHA256 digest to be signed",
                      "example": "4LedqdTMbBbnZF9YjfxVkIZYbG51lgYuubiNp9EqjS4="
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/issue": {
      "post": {
        "summary": "Issues a signature",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "bytes",
                  "issuerId"
                ],
                "type": "object",
                "properties": {
                  "bytes": {
                    "type": "string",
                    "description": "Base64 encoded SHA256 digest to be signed",
                    "example": "4LedqdTMbBbnZF9YjfxVkIZYbG51lgYuubiNp9EqjS4="
                  },
                  "issuerId": {
                    "type": "string",
                    "description": "Issuer ID used to identify institution or staff member requesting issuance",
                    "example": "8d51fa75-b98e-4d8f-98f1-dee5d471a450"
                  },
                  "auditLog": {
                    "type": "string",
                    "description": "Optional audit log entry",
                    "example": "[2023-04-01T12:00:00Z] School #314, Staff #331.A"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cms": {
                      "$ref": "#/components/schemas/Cms"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/merge/pdf": {
      "post": {
        "summary": "Produces a signed PDF by merging the original PDF with its signature.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "bytes",
                  "signingTimestamp",
                  "cms"
                ],
                "type": "object",
                "properties": {
                  "bytes": {
                    "type": "string",
                    "description": "Base64 encoded original PDF used to generate the digest to be signed",
                    "example": "JVBERi0xLjEKJcKlwrHDqwoKMSAwIG9iagogIDw8IC9UeXBlIC9DYXRhbG9nCiAgICAgL1BhZ2VzIDIgMCBSCiAgPj4KZW5kb2JqCgoyIDAgb2JqCiAgPDwgL1R5cGUgL1BhZ2VzCiAgICAgL0tpZHMgWzMgMCBSXQogICAgIC9Db3VudCAxCiAgICAgL01lZGlhQm94IFswIDAgMzAwIDE0NF0KICA+PgplbmRvYmoKCjMgMCBvYmoKICA8PCAgL1R5cGUgL1BhZ2UKICAgICAgL1BhcmVudCAyIDAgUgogICAgICAvUmVzb3VyY2VzCiAgICAgICA8PCAvRm9udAogICAgICAgICAgIDw8IC9GMQogICAgICAgICAgICAgICA8PCAvVHlwZSAvRm9udAogICAgICAgICAgICAgICAgICAvU3VidHlwZSAvVHlwZTEKICAgICAgICAgICAgICAgICAgL0Jhc2VGb250IC9UaW1lcy1Sb21hbgogICAgICAgICAgICAgICA+PgogICAgICAgICAgID4+CiAgICAgICA+PgogICAgICAvQ29udGVudHMgNCAwIFIKICA+PgplbmRvYmoKCjQgMCBvYmoKICA8PCAvTGVuZ3RoIDU1ID4+CnN0cmVhbQogIEJUCiAgICAvRjEgMTggVGYKICAgIDAgMCBUZAogICAgKEhlbGxvIFdvcmxkKSBUagogIEVUCmVuZHN0cmVhbQplbmRvYmoKCnhyZWYKMCA1CjAwMDAwMDAwMDAgNjU1MzUgZiAKMDAwMDAwMDAxOCAwMDAwMCBuIAowMDAwMDAwMDc3IDAwMDAwIG4gCjAwMDAwMDAxNzggMDAwMDAgbiAKMDAwMDAwMDQ1NyAwMDAwMCBuIAp0cmFpbGVyCiAgPDwgIC9Sb290IDEgMCBSCiAgICAgIC9TaXplIDUKICA+PgpzdGFydHhyZWYKNTY1CiUlRU9GCg=="
                  },
                  "signingTimestamp": {
                    "type": "number",
                    "description": "Original signing timestamp used to generate the digest to be signed",
                    "example": 1676906127623
                  },
                  "cms": {
                    "$ref": "#/components/schemas/Cms"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bytes": {
                      "type": "string",
                      "description": "Base64 encoded signed PDF"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/validate/pdf": {
      "post": {
        "summary": "Validates a signed PDF's signature",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "bytes"
                ],
                "type": "object",
                "properties": {
                  "bytes": {
                    "type": "string",
                    "description": "Base64 encoded signed PDF"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean",
                      "description": "Overall validation result. Returns true if and only if all aspects of the validation are successful."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cms": {
        "type": "string",
        "description": "Base64 encoded signature in CMS format",
        "example": "MIAGCSqGSIb3DQEHAqCAMIACAQExDTALBglghkgBZQMEAgEwgAYJKoZIhvcNAQcBAACggDCCAdQwggF7oAMCAQICBGOAtQwwCgYIKoZIzj0EAwIwZjELMAkGA1UEBhMCREUxFTATBgNVBAgMDEJ1bmRlc2xhbmQgQTEdMBsGA1UECgwUQklSRCBXUDcgREVWRUxPUE1FTlQxITAfBgNVBAMMGExhbmRlc2ViZW5lIEJ1bmRlc2xhbmQgQTAeFw0yMjExMjUxMjI5MDBaFw0yMzExMjUxMjI5MDBaMG4xCzAJBgNVBAYTAkRFMRUwEwYDVQQIDAxCdW5kZXNsYW5kIEExEDAOBgNVBAcMB1N0YWR0IEExHTAbBgNVBAoMFEJJUkQgV1A3IERFVkVMT1BNRU5UMRcwFQYDVQQDDA5TY2h1bGUgU3RhZHQgQTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABBzIem/FyzQBZEbK7kUnN3SpwWCIGvwfWPugt6BRS1zpR5pLAFvJP4o8h/GnDQ6g+BP4lJYi5GZxjSkvXhxCM5KjDzANMAsGA1UdDwQEAwIHgDAKBggqhkjOPQQDAgNHADBEAiB1bbBVeZqIRVYF8/djpRIP/wIHACU8XKIadFPCzFQKegIgEaVHaxcS6sk2XR0+5QO+IIHUfUsbp+YBQ4VxfOowZO4wggH9MIIBpKADAgECAgRjgLTPMAoGCCqGSM49BAMCMEMxCzAJBgNVBAYTAkRFMR0wGwYDVQQKDBRCSVJEIFdQNyBERVZFTE9QTUVOVDEVMBMGA1UEAwwMQnVuZGVzLUViZW5lMB4XDTIyMTEyNTEyMjc1OVoXDTIzMTEyNTEyMjc1OVowZjELMAkGA1UEBhMCREUxFTATBgNVBAgMDEJ1bmRlc2xhbmQgQTEdMBsGA1UECgwUQklSRCBXUDcgREVWRUxPUE1FTlQxITAfBgNVBAMMGExhbmRlc2ViZW5lIEJ1bmRlc2xhbmQgQTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABJrL2e9+Q/JBA3BeAknXRLXxmRpVTP+pA/YXi/afLeROUFv2+kiwBfhJIAuDMJK5yHdu2kP6AyWQr9j23ZnjRBSjYzBhMA8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgEGMB8GA1UdIwQYMBaAFLUSJqUiZOqPqBp1RwB07mnIi9OnMB0GA1UdDgQWBBSWuZvZGzgd7Y7bWXfRiIEEDZvkaTAKBggqhkjOPQQDAgNHADBEAiBCovXLBkA5EFvap0Oi+bL1UFMtaHWgba0TpJq169GR0QIgLBhEcX/ilY7QDWmKcWAVUelfPBHD4lLCpUikX0iMBxQwggHbMIIBgaADAgECAgRjgLSUMAoGCCqGSM49BAMCMEMxCzAJBgNVBAYTAkRFMR0wGwYDVQQKDBRCSVJEIFdQNyBERVZFTE9QTUVOVDEVMBMGA1UEAwwMQnVuZGVzLUViZW5lMB4XDTIyMTEyNTEyMjcwMFoXDTIzMTEyNTEyMjcwMFowQzELMAkGA1UEBhMCREUxHTAbBgNVBAoMFEJJUkQgV1A3IERFVkVMT1BNRU5UMRUwEwYDVQQDDAxCdW5kZXMtRWJlbmUwWTATBgcqhkjOPQIBBggqhkjOPQMBBwNCAATeFUPWTFaX5pOcS5eiWCaV0gyvTZKHZ4rQ3vy6YkQnC799TEOqkUoaYEht1f8FxNJMAqdMvDgwbAdjLAh/J+VYo2MwYTAPBgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIBBjAfBgNVHSMEGDAWgBS1EialImTqj6gadUcAdO5pyIvTpzAdBgNVHQ4EFgQUtRImpSJk6o+oGnVHAHTuaciL06cwCgYIKoZIzj0EAwIDSAAwRQIhAJCCMLy5jRGDb34ip/NE9ncGuQKpkZ6ehTd8EbkB6oudAiBD29O7VEa0xBkVWDwFRw4U6yoc7pVh5TSlQurm4bVSuAAAMYIb5TCCG+ECAQEwbjBmMQswCQYDVQQGEwJERTEVMBMGA1UECAwMQnVuZGVzbGFuZCBBMR0wGwYDVQQKDBRCSVJEIFdQNyBERVZFTE9QTUVOVDEhMB8GA1UEAwwYTGFuZGVzZWJlbmUgQnVuZGVzbGFuZCBBAgRjgLUMMAsGCWCGSAFlAwQCAaCB/TAYBgkqhkiG9w0BCQMxCwYJKoZIhvcNAQcBMC8GCSqGSIb3DQEJBDEiBCAejdkTrl2MWC3XcRa9ZrCinBkgBEfgmON0IDUUHjZOSjCBrwYLKoZIhvcNAQkQAi8xgZ8wgZwwgZkwgZYEIFFD+LMVvYlfAMD8bbaKUxUi0Q66P6g2Wgk0xBbJWu4QMHIwaqRoMGYxCzAJBgNVBAYTAkRFMRUwEwYDVQQIDAxCdW5kZXNsYW5kIEExHTAbBgNVBAoMFEJJUkQgV1A3IERFVkVMT1BNRU5UMSEwHwYDVQQDDBhMYW5kZXNlYmVuZSBCdW5kZXNsYW5kIEECBGOAtQwwCgYIKoZIzj0EAwIESDBGAiEAsBxyxu/1YkjE6fivZFu+b/eA1WE/AZlW/iL4W4bGK+MCIQDJQ/FXU80dHDN4icMXgVaZxWzAGzJ1lB+cDuYVQpzlIaGCGgcwghoDBgsqhkiG9w0BCRACDjGCGfIwghnuBgkqhkiG9w0BBwKgghnfMIIZ2wIBAzEPMA0GCWCGSAFlAwQCAQUAMIGEBgsqhkiG9w0BCRABBKB1BHMwcQIBAQYMKwYBBAGBrSGCLBYBMDEwDQYJYIZIAWUDBAIBBQAEIEIwaMxVKdRiiP7ein23qJUI1LXEWvhVC1xh8dpZmejlAhQdOnUdyU3GHhJM73ouyy2CUbK+mhgPMjAyMzAyMjAxNTMzMDhaAgQTZ7xMoIIVmDCCA8MwggKroAMCAQICAQEwDQYJKoZIhvcNAQELBQAwgYIxCzAJBgNVBAYTAkRFMSswKQYDVQQKDCJULVN5c3RlbXMgRW50ZXJwcmlzZSBTZXJ2aWNlcyBHbWJIMR8wHQYDVQQLDBZULVN5c3RlbXMgVHJ1c3QgQ2VudGVyMSUwIwYDVQQDDBxULVRlbGVTZWMgR2xvYmFsUm9vdCBDbGFzcyAyMB4XDTA4MTAwMTEwNDAxNFoXDTMzMTAwMTIzNTk1OVowgYIxCzAJBgNVBAYTAkRFMSswKQYDVQQKDCJULVN5c3RlbXMgRW50ZXJwcmlzZSBTZXJ2aWNlcyBHbWJIMR8wHQYDVQQLDBZULVN5c3RlbXMgVHJ1c3QgQ2VudGVyMSUwIwYDVQQDDBxULVRlbGVTZWMgR2xvYmFsUm9vdCBDbGFzcyAyMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAql/aG1/oc5Hl2lz0ouZH5fNoVWAFHQKks5tZ8x6KrzSt/A3C2UgZ7mmPySD8IaoHGe2wXKxlx1/tAnx7fC0b1rq5gMIYghaE+mawCMZUI4HkzblJP/ZPbjdIKDgPxb7naHD9OZdN0seYkVCqxESzI305R+lSYtYSk163MZZCBft2px6j9cL86XrFbKlxT+rLeLxgr8fe9NnLvn4zpW6Ug/A0+iGr6o5yoD+k3jBb74ZNapVbQ0SoEBUc5QFXxZjx5gYokaogxbdTJlFDsgsRlVjhwA922cCNfIHzcnCeb/4ajtlfNcaybzR8vkhP4lo519ideJ6fhj4DXhmLRKLVxwIDAQABo0IwQDAPBgNVHRMBAf8EBTADAQH/MA4GA1UdDwEB/wQEAwIBBjAdBgNVHQ4EFgQUv1kgNgB5oKAia4zV8mHSuCzLgkowDQYJKoZIhvcNAQELBQADggEBADEDomELH3TocjbGbflNnvoiqOGBVs/Nu5/qq5EZOK+qfBVN87ajjaX0jvZEqafoIZWtPgBiFojwArr8YSPmM5swems2YnutBCOEWGXi2yuK5yVTN2JTX7zaAWIpoqYnceY6In7Bbx2VcCBKBzTf6v8VgOW613rYW3V8BXopR35AqDETd81AO7RRR3ouEeNHEd6dZtCL1VRm+oNV6nzCKYkb6W+zzuIFhMkvPniFYm7JX8F4Y3RYwEgYDJk566TMGrV5Wo0VnNgUDfZ6B1fHIoMFLTybJSY9GLOpQ3zIyKtkjw6jv5wbnTDb2tAZLqo88fszgHbkza0ZTwUnjhOhbsIwggUSMIID+qADAgECAgkA4wvV+K8l2YEwDQYJKoZIhvcNAQELBQAwgYIxCzAJBgNVBAYTAkRFMSswKQYDVQQKDCJULVN5c3RlbXMgRW50ZXJwcmlzZSBTZXJ2aWNlcyBHbWJIMR8wHQYDVQQLDBZULVN5c3RlbXMgVHJ1c3QgQ2VudGVyMSUwIwYDVQQDDBxULVRlbGVTZWMgR2xvYmFsUm9vdCBDbGFzcyAyMB4XDTE2MDIyMjEzMzgyMloXDTMxMDIyMjIzNTk1OVowgZUxCzAJBgNVBAYTAkRFMUUwQwYDVQQKEzxWZXJlaW4genVyIEZvZXJkZXJ1bmcgZWluZXMgRGV1dHNjaGVuIEZvcnNjaHVuZ3NuZXR6ZXMgZS4gVi4xEDAOBgNVBAsTB0RGTi1QS0kxLTArBgNVBAMTJERGTi1WZXJlaW4gQ2VydGlmaWNhdGlvbiBBdXRob3JpdHkgMjCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMtg1/9moUHN0vqHl4pzq5lN6mc5WqFggEcVToyVsuXPztNXS43O+FZsFVV2B+pG/cgDRWM+cNSrVICxI5y+NyipCf8FXRgPxJiZN7Mg9mZ4F4fCnQ7MSjLnFp2uDo0peQcAIFTcFV9Kltd4tjTTwXS1nem/wHdN6r1ZB+BaL2w8pQDcNb1lDY9/Mm3yWmpLYgHurDg0WUU2SQXaeMpqbVvAgWsRzNI8qIv4cRrKO+KA3Ra0Z3qLNupOkSk9s1FcragMvp0049ENF4N1xDkesJQLEvHVaY4l9Lg9K7/AjsMeO6W/VRCrKq4Xl14zzsjz9AkH4wKGMUZrAcUQDBHHWekCAwEAAaOCAXQwggFwMA4GA1UdDwEB/wQEAwIBBjAdBgNVHQ4EFgQUk+PYMiba1fFKpZFK4OpL4qIMz+EwHwYDVR0jBBgwFoAUv1kgNgB5oKAia4zV8mHSuCzLgkowEgYDVR0TAQH/BAgwBgEB/wIBAjAzBgNVHSAELDAqMA8GDSsGAQQBga0hgiwBAQQwDQYLKwYBBAGBrSGCLB4wCAYGZ4EMAQICMEwGA1UdHwRFMEMwQaA/oD2GO2h0dHA6Ly9wa2kwMzM2LnRlbGVzZWMuZGUvcmwvVGVsZVNlY19HbG9iYWxSb290X0NsYXNzXzIuY3JsMIGGBggrBgEFBQcBAQR6MHgwLAYIKwYBBQUHMAGGIGh0dHA6Ly9vY3NwMDMzNi50ZWxlc2VjLmRlL29jc3ByMEgGCCsGAQUFBzAChjxodHRwOi8vcGtpMDMzNi50ZWxlc2VjLmRlL2NydC9UZWxlU2VjX0dsb2JhbFJvb3RfQ2xhc3NfMi5jZXIwDQYJKoZIhvcNAQELBQADggEBAIcL/z4Cm2XIVi3WO5qYi3FP2ropqiH5Ri71sqQPrhE4eTizDnS6dl2e6BiClmLbTDPo3flq3zK9LExHYFV/53RrtCyD2HlrtrdNUAtmB7Xts5et6u5/MOaZ/SLick0+hFvu+c+Z6n/XUjkurJgARH5pO7917tALOxrN5fcPImxHhPalR6D90Bo0fa3SPXez7vTXTf/D6OWST1k+kEcQSrCFWMBvf/iu7QhCnh7U3xQuTY+8npTD5+32GPg8SecmqKc22CzeIs2LgtjZeOJVEqM7h0S2EQvVDFKvaYwPBt/QolOLV5h7z/0HJPT8vcP9SpIClxvyt7bPZYoaorVyGTkwggWsMIIElKADAgECAgcbY7rQHiw9MA0GCSqGSIb3DQEBCwUAMIGVMQswCQYDVQQGEwJERTFFMEMGA1UEChM8VmVyZWluIHp1ciBGb2VyZGVydW5nIGVpbmVzIERldXRzY2hlbiBGb3JzY2h1bmdzbmV0emVzIGUuIFYuMRAwDgYDVQQLEwdERk4tUEtJMS0wKwYDVQQDEyRERk4tVmVyZWluIENlcnRpZmljYXRpb24gQXV0aG9yaXR5IDIwHhcNMTYwNTI0MTEzODQwWhcNMzEwMjIyMjM1OTU5WjCBjTELMAkGA1UEBhMCREUxRTBDBgNVBAoMPFZlcmVpbiB6dXIgRm9lcmRlcnVuZyBlaW5lcyBEZXV0c2NoZW4gRm9yc2NodW5nc25ldHplcyBlLiBWLjEQMA4GA1UECwwHREZOLVBLSTElMCMGA1UEAwwcREZOLVZlcmVpbiBHbG9iYWwgSXNzdWluZyBDQTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJ07eRxH3h+Gy8Zp1xCeOdfZojDbchwFfylfS2jxrRnWTOFrG7ELf6Gr4HuLi9gtzm6IOhDuV+UefwRRNuu6cG1joL6WLkDh0YNMZj0cZGnlm6Stcq5oOVGHecwX064vXWNxSzl660Knl5BpBb+Q/6RAcL0D57+eGIgfn5mITQ5HjUhfZZkQ0tkqSe3BuS0dnxLLFdM/fx5ULzquk1enfnjK1UriGuXtQX1TX8izKvWKMKztFwUkP7agCwf9TRqaA1KgNpzeJIdl5Of6x5ZzJBTN0OgbaJ4YWa52fvfRCng8h0uwN89Tyjo4EPPLR22MZD08WkVKusqAfLjz56dMTM0CAwEAAaOCAgUwggIBMBIGA1UdEwEB/wQIMAYBAf8CAQEwDgYDVR0PAQH/BAQDAgEGMCkGA1UdIAQiMCAwDQYLKwYBBAGBrSGCLB4wDwYNKwYBBAGBrSGCLAEBBDAdBgNVHQ4EFgQUazqYi/nyU4na4K2yMh4JH+iqO3QwHwYDVR0jBBgwFoAUk+PYMiba1fFKpZFK4OpL4qIMz+EwgY8GA1UdHwSBhzCBhDBAoD6gPIY6aHR0cDovL2NkcDEucGNhLmRmbi5kZS9nbG9iYWwtcm9vdC1nMi1jYS9wdWIvY3JsL2NhY3JsLmNybDBAoD6gPIY6aHR0cDovL2NkcDIucGNhLmRmbi5kZS9nbG9iYWwtcm9vdC1nMi1jYS9wdWIvY3JsL2NhY3JsLmNybDCB3QYIKwYBBQUHAQEEgdAwgc0wMwYIKwYBBQUHMAGGJ2h0dHA6Ly9vY3NwLnBjYS5kZm4uZGUvT0NTUC1TZXJ2ZXIvT0NTUDBKBggrBgEFBQcwAoY+aHR0cDovL2NkcDEucGNhLmRmbi5kZS9nbG9iYWwtcm9vdC1nMi1jYS9wdWIvY2FjZXJ0L2NhY2VydC5jcnQwSgYIKwYBBQUHMAKGPmh0dHA6Ly9jZHAyLnBjYS5kZm4uZGUvZ2xvYmFsLXJvb3QtZzItY2EvcHViL2NhY2VydC9jYWNlcnQuY3J0MA0GCSqGSIb3DQEBCwUAA4IBAQCBeEWkTqR/DlXwCbFqPnjMaDWpHPOVnj/z+N9rOHeJLI21rT7H8pTNoAauusyosa0zCLYkhmI2THhuUPDVbmCNT1IxQ5dGdfBi5G5mUcFCMWdQ5UnnOR7Ln8qGSN4IFP8VSytmm6A4nwDO/afr0X9XLchMX9wQEZc+lgQCXISoKTlslPwQkgZ7nu7YRrQbtQMMONncsKk/cQYLsgMHM8KNSGMlJTx6e1du94oFOO+4oK4v9NsH1VuEGMGpuEvObJAaguS5Pfp38dIfMwK/U+d2+dwmJUFvL6Yb+qQTkPp8ftkLYF3sv8pBoGH7EUkp2KgtdRXYShjqFu9VNCIaE40GMIIHBzCCBe+gAwIBAgIMJ2Y14U/n46aIxvOdMA0GCSqGSIb3DQEBCwUAMIGNMQswCQYDVQQGEwJERTFFMEMGA1UECgw8VmVyZWluIHp1ciBGb2VyZGVydW5nIGVpbmVzIERldXRzY2hlbiBGb3JzY2h1bmdzbmV0emVzIGUuIFYuMRAwDgYDVQQLDAdERk4tUEtJMSUwIwYDVQQDDBxERk4tVmVyZWluIEdsb2JhbCBJc3N1aW5nIENBMB4XDTIyMTAxMjEyNTEzMloXDTI1MTAxMTEyNTEzMlowgcsxCzAJBgNVBAYTAkRFMQ8wDQYDVQQIDAZCZXJsaW4xDzANBgNVBAcMBkJlcmxpbjFFMEMGA1UECgw8VmVyZWluIHp1ciBGb2VyZGVydW5nIGVpbmVzIERldXRzY2hlbiBGb3JzY2h1bmdzbmV0emVzIGUuIFYuMRkwFwYDVQQLDBBHZXNjaGFlZnRzc3RlbGxlMRkwFwYDVQRBDBBaZWl0c3RlbXBlbCAyMDIyMR0wGwYDVQQDDBRQTjogWmVpdHN0ZW1wZWwgMjAyMjCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBALRbHmgZl1ctJmwR6WdW2yNe2nC4EGNwNjqBxF5dxZTY5hqwuW4N2lCj09cf1Riyd+lab/Dwoy67zcy+0L6ui+mYhx417BqMDSTKk/5I4CWvWzetHLAp33Wx+DWHMiaflQ2jzajo9+jZLrkWv5BMIwI8lEPEmXXXlXvnncowBseZiV/rzBj/Hs2iQgK6ae49lhX0ybLllftsS4hHcl3rqN9tDPM8mb+U9DVKnyskARoZM4EpdpqBW/wpQKlnN24J5/cZ9xC2/9tKFLjBN+y10xg2teJi7CSU9XNofyroRHQcANDnoRPKxQmRw/4fSp2j65NmmKQxDKoZ80npd+rFP8C2cSlFLtUN97ViLIiWZojGpe3yFQdlq+gpmi2zAwZ7y+Y49DMpLq7TXUS0onGQaLH4Ca4buowU3TI2E4dGpa4DXLWebIxRzRD7W/N3BUFx7bYqWs/R5nQjk5DSrVNcgjTE/Z1EllGCSAijgsdFzDDwywUBeKDdz52E3RJzQnfbXh/lj1HEL8gPWOg7Cch2hfZuPojJ/uv+D5usbatJv2dtXn7UOhVDQf+A4Qc1wE5sS/v0BW2AuksqDnrRlk1zcV4JU02fl+6Q+/J32oTQ1ihuOo7nUwk5GthPJfjbtu9x6BM46nLUeJQtCKTz5SFbxTIyiG9R+9vCF4KbSrVmLtpbAgMBAAGjggIlMIICITA+BgNVHSAENzA1MA8GDSsGAQQBga0hgiwBAQQwEAYOKwYBBAGBrSGCLAEBBAowEAYOKwYBBAGBrSGCLAIBBAowCQYDVR0TBAIwADAOBgNVHQ8BAf8EBAMCB4AwFgYDVR0lAQH/BAwwCgYIKwYBBQUHAwgwHQYDVR0OBBYEFJdxSeBnQ7s8c+lj38OyeuEyCrKTMB8GA1UdIwQYMBaAFGs6mIv58lOJ2uCtsjIeCR/oqjt0MIGNBgNVHR8EgYUwgYIwP6A9oDuGOWh0dHA6Ly9jZHAxLnBjYS5kZm4uZGUvZGZuLWNhLWdsb2JhbC1nMi9wdWIvY3JsL2NhY3JsLmNybDA/oD2gO4Y5aHR0cDovL2NkcDIucGNhLmRmbi5kZS9kZm4tY2EtZ2xvYmFsLWcyL3B1Yi9jcmwvY2FjcmwuY3JsMIHbBggrBgEFBQcBAQSBzjCByzAzBggrBgEFBQcwAYYnaHR0cDovL29jc3AucGNhLmRmbi5kZS9PQ1NQLVNlcnZlci9PQ1NQMEkGCCsGAQUFBzAChj1odHRwOi8vY2RwMS5wY2EuZGZuLmRlL2Rmbi1jYS1nbG9iYWwtZzIvcHViL2NhY2VydC9jYWNlcnQuY3J0MEkGCCsGAQUFBzAChj1odHRwOi8vY2RwMi5wY2EuZGZuLmRlL2Rmbi1jYS1nbG9iYWwtZzIvcHViL2NhY2VydC9jYWNlcnQuY3J0MA0GCSqGSIb3DQEBCwUAA4IBAQBI5ptg73IfTRFPgi4/mb9bXMt2o/hjJSx1Cnb60b1T3+4Nsd8BCtHce0lGTNWlqf5tJE7ZW0SMc+z4SuST0bnB6eYfC5eQTP93Yl8PBwy22JOjVaDnu95jw66TM3hFpQXBoG/0xoLV1bAhMKFcBE1CKFTTdSbNXKvUhxtFeHp4GHftWxDc1zsGngDSf8egVbVCOuDBMccSmYIrkfjDUeGuPoy8u2Y1Bhp9SCNvCcMnBeL0xisYvGXCDPnPjqH31E2LPJSLMcv8BuJUdlsh0UIUJRPzM6HXpKQOMlzwZuHbBIP+XCEoSGpHAMiP1O3kaVkOT52Xrr/A6Sq4q5zXboS2MYIDoDCCA5wCAQEwgZ4wgY0xCzAJBgNVBAYTAkRFMUUwQwYDVQQKDDxWZXJlaW4genVyIEZvZXJkZXJ1bmcgZWluZXMgRGV1dHNjaGVuIEZvcnNjaHVuZ3NuZXR6ZXMgZS4gVi4xEDAOBgNVBAsMB0RGTi1QS0kxJTAjBgNVBAMMHERGTi1WZXJlaW4gR2xvYmFsIElzc3VpbmcgQ0ECDCdmNeFP5+OmiMbznTANBglghkgBZQMEAgEFAKCB0zAaBgkqhkiG9w0BCQMxDQYLKoZIhvcNAQkQAQQwHAYJKoZIhvcNAQkFMQ8XDTIzMDIyMDE1MzMwOFowLQYJKoZIhvcNAQk0MSAwHjANBglghkgBZQMEAgEFAKENBgkqhkiG9w0BAQsFADAvBgkqhkiG9w0BCQQxIgQgPdr8dUMLPPSb6srGxsX2wrTcEqQ9TjC52A1Xf/yEP0YwNwYLKoZIhvcNAQkQAi8xKDAmMCQwIgQgZI1/BlmSjvGCC4uqKTQ3nhnbS0VoTZTH0JUeD9dUMa0wDQYJKoZIhvcNAQELBQAEggIAO6+zRAQPjqDNzu/BZgOGH/zvV/oDUVXxARMRr9bXKU8rsEkfrMdPgeFeH/q7hMVfgTTzgzDJnsi5ZH6WjY/JOitvnHE1VWb33LkcpeGSY5gn7qXziIfzxmRnYWihfucVpfTkRXmhVXWPQmjjG/1j8lI1DsYFW0u87nHvSRy7srh5F+5vWix5QWZAwDAt8rmABks9d2cpUzf3hYPVkVOxYUV8yn7zAWxRm9pheYhul833pCO4ec1yL3qQPByASJ909mB234xXGs92tLoEKtsMZSkNczENfOO4h+MuPT5Knncz2RPkQshQdR9Yh3W/aCDGnXrI2ootBHwJn3QO6gPVofB5IwVjesn/zbW1nrMDgfLsd0u2I9EnhsfSqV4OxpA3pMw+Uv3cGjiwJhMDD7zpoySc3aSTfldpKJPRD1NAStEWePIU9D3CTtXNo8sB+Zw+NUotLOvUn9BCzI+UBwbGPsCu3iHC/Oyphj1EguUdDkCnuNtVRQ9Sk04E+b1Ii6i7dvUyyu5iF6pGwIhVHcNK32aYzVfctFLJ2/b1iuDBnaPk2ty3SxnxfA4sgkcRULo870ZlGkNrpMVUICLAnPELm0eZo+Olz0fM3m2xpfQQeWcAuG9SVCm0hbNZenuWBzaexu9EnPWqbB8WZ4CxIuC8oVxioXk7ZCYJ05ZamkzK5ygAAAAAAAA="
      }
    }
  }
}